added few modification #3

+++ src/db/migrations/20190220224645-create-product.js
@@ -30,7 +30,7 @@ module.exports = {
 				type: Sequelize.STRING(150),
 			},
 			display: {
-				type: Sequelize.SMALLINT(6),
+				type: Sequelize.INTEGER,
 			},
 			// category_id: {
 			// 	type: Sequelize.INTEGER,
